+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add `PointerScope` to manage more easily the resources of a group of `Pointer` objects
  * Fix `Parser` failing on `const void*&` or similar function arguments, and on constructors of class templates
  * Add `Info.skipDefaults` to have the `Parser` ignore default function arguments and prevent method overloading
  * Accelerate copy and extraction of resources by using larger buffers for file operations